pyproject.toml
[tool.poetry]
name = "Inventory-Management-system"
version = "0.1.0"
description = "Backend system for an eCommerce store"
authors = ["ParzivalXII <kayodeibrahim97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "~0.116.0"
uvicorn = {extras = ["standard"], version = "~0.35.0"}
sqlmodel = "0.0.14"
python-jose = {extras = ["cryptography"], version = "~3.5.0"}
passlib = {extras = ["bcrypt"], version = "~1.7.4"}
redis = "~5.0.4"
pydantic = "~2.11.0"
pydantic-email = "~2.11.0"
pydantic-settings = "~2.0.3"
python-multipart = "~0.0.9"
python-dotenv = "~1.0.1"
slowapi = "~0.1.9"

[tool.poetry.dev-dependencies]
pytest = "~8.2.2"
httpx = "~0.27.0"
pytest-asyncio = "~0.23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"